metricsbackend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
bcrypt
PyJWT
pandas
numpy
scikit-learn
joblib
pytest
python-dateutil
fastapi
uvicorn
prophet
matplotlib
seaborn
plotly
gunicorn 